+++ test/components/text_input_test.exs
@@ -195,11 +195,8 @@ defmodule PrimerLive.TestComponents.TextInputTest do
     }
 
     assert rendered_to_string(~H"""
-           <.text_input caption={
-             fn ->
-               "Caption"
-             end
-           } />
+           <.text_input caption={fn -> "Caption" end} />
+           <.text_input caption={fn -> "Caption" end} is_form_control />
            <.text_input caption={
              fn field_state ->
                if !field_state.valid?,
@@ -212,6 +209,11 @@ defmodule PrimerLive.TestComponents.TextInputTest do
              """
              <input class="FormControl-input FormControl-medium" type="text" />
              <div class="FormControl-caption">Caption</div>
+             <div class="FormControl">
+             <div class="form-group-header"><label class="FormControl-label"></label></div><input
+             class="FormControl-input FormControl-medium" type="text" />
+             <div class="FormControl-caption">Caption</div>
+             </div>
              <input class="FormControl-input FormControl-medium" type="text" />
              """
              |> format_html()
@@ -475,37 +477,86 @@ defmodule PrimerLive.TestComponents.TextInputTest do
   end
 
   test "Attribute: classes" do
-    assigns = %{}
+    assigns = %{
+      classes: %{
+        input: "input-x",
+        input_group: "input_group-x",
+        input_group_button: "input_group_button-x",
+        validation_message: "validation_message-x",
+        input_wrap: "input_wrap-x",
+        caption: "caption-x"
+      },
+      form_control_attrs: %{
+        classes: %{
+          control: "control-x",
+          group: "group-x",
+          header: "header-x",
+          label: "label-x"
+        }
+      },
+      class: "my-text-input",
+      form: %{@default_form | source: @error_changeset}
+    }
 
     assert rendered_to_string(~H"""
            <.text_input
-             classes={
-               %{
-                 form_control: "form_control-x",
-                 input: "input-x",
-                 input_group: "input_group-x",
-                 input_group_button: "input_group_button-x",
-                 validation_message: "validation_message-x",
-                 input_wrap: "input_wrap-x"
-               }
-             }
-             class="my-text-input"
+             classes={@classes}
+             form_control={@form_control_attrs}
+             class={@class}
+             caption={fn -> "Caption" end}
+             form={@form}
+             field={:first_name}
            >
              <:group_button>
                <.button>Send</.button>
              </:group_button>
+           </.text_input>
+
+           <.text_input classes={@classes} form_control={@form_control_attrs} class={@class}>
              <:leading_visual class="my-leading-visual">
                <.octicon name="mail-16" />
              </:leading_visual>
            </.text_input>
+
+           <.text_input classes={@classes} form_control={@form_control_attrs} class={@class}>
+             <:trailing_action>
+               <.button is_icon_only aria-label="Clear">
+                 <.octicon name="x-16" />
+               </.button>
+             </:trailing_action>
+           </.text_input>
            """)
            |> format_html() ==
              """
+             <div class="FormControl group-x control-x pl-invalid">
+             <div class="form-group-header header-x"><label class="FormControl-label label-x" for="user_first_name">First
+             name</label></div>
              <div class="input-group input_group-x">
-             <input class="FormControl-input FormControl-medium input-x my-text-input" type="text" />
-             <span class="input-group-button input_group_button-x">
-             <button class="btn" type="button">Send</button>
-             </span>
+             <div class="pl-invalid" phx-feedback-for="user[first_name]"><input aria-describedby="user_first_name-validation"
+             class="FormControl-input FormControl-medium input-x my-text-input" id="user_first_name" invalid=""
+             name="user[first_name]" type="text" value="" /></div><span
+             class="input-group-button input_group_button-x"><button class="btn" type="button">Send</button></span>
+             </div>
+             <div class="FormControl-inlineValidation FormControl-inlineValidation--error validation_message-x"
+             id="user_first_name-validation" phx-feedback-for="user[first_name]"><svg class="octicon"
+             xmlns="http://www.w3.org/2000/svg" width="12" height="12"
+             viewBox="0 0 12 12">STRIPPED_SVG_PATHS</svg><span>can&#39;t be blank</span></div>
+             <div class="FormControl-caption caption-x">Caption</div>
+             </div>
+             <div class="FormControl group-x control-x">
+             <div class="form-group-header header-x"><label class="FormControl-label label-x"></label></div>
+             <div class="FormControl-input-wrap FormControl-input-wrap--leadingVisual input_wrap-x"><span
+             class="FormControl-input-leadingVisualWrap"><svg class="octicon" xmlns="http://www.w3.org/2000/svg" width="16"
+             height="16" viewBox="0 0 16 16">STRIPPED_SVG_PATHS</svg></span><input
+             class="FormControl-input FormControl-medium input-x my-text-input" type="text" /></div>
+             </div>
+             <div class="FormControl group-x control-x">
+             <div class="form-group-header header-x"><label class="FormControl-label label-x"></label></div>
+             <div class="FormControl-input-wrap FormControl-input-wrap--trailingAction input_wrap-x"><input
+             class="FormControl-input FormControl-medium input-x my-text-input" type="text" /><span
+             class="FormControl-input-trailingAction"><button aria-label="Clear" class="btn-octicon" type="button"><svg
+             class="octicon" xmlns="http://www.w3.org/2000/svg" width="16" height="16"
+             viewBox="0 0 16 16">STRIPPED_SVG_PATHS</svg></button></span></div>
              </div>
              """
              |> format_html()